Modified all timesteps and durations to be represented by int or Integer rather than long or Long.
Added an interface for Semi-Markov decision processes (SMDP.java).

+++ src/com/github/kingtim1/jmdp/DeterministicPolicy.java
@@ -45,7 +45,7 @@ public abstract class DeterministicPolicy<S, A> implements
 	}
 
 	@Override
-	public A policy(S state, Long timestep) {
+	public A policy(S state, Integer timestep) {
 		return policy(state);
 	}
